requirements.txt
Django==3.2.13
easyocr==1.7.1
joblib==1.1.0
numpy==1.26.4
openai==1.30.1
pandas==1.3.4
Pillow==10.3.0
protobuf==3.20.0
pytesseract==0.3.10
scikit_learn==1.4.2
tensorflow==2.8.0
tensorflow_gpu==2.10.0
tensorflow_intel==2.16.1
